refactor test files and update coverage settings
