+++ test/App.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import App from './App';
+import App from '../src/App';
 
 describe('App Component', () => {
   it('renders the Header component', () => {
