+++ test/components/Footer.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import Footer from './Footer';
+import Footer from '../../src/components/Footer';
 
 describe('Footer Component', () => {
   it('renders the developer information', () => {
